intermediate commit
git-svn-id: https://max-pc/svn/MaxSVNrepository/Astro/trunk/Utils@2167 5b3607db-4df2-0244-8a27-3f6f91efc260

+++ MakeStack/CalcThread.pas
@@ -65,8 +65,8 @@ type
     Naxis: TIntArray;
     BScale: Double;
     BZero: Double;
-    StartOfImage: Int64;
-    ImageMemSize: PtrUInt; // Padded!
+    StartOfImage: Integer;
+    ImageMemSize: PtrUInt; // Padded to FITSRecordLen!
     ObjectName: string;
     Telescope: string;
     Instrument: string;
